Break code into modules
Only a refactor, no substantive changes.
